feat: move showmore modal out into own comp
